from long to int in SPINRGBOOT

+++ Back/Events/src/main/java/com/example/supplier/Controller/EventController.java
@@ -30,7 +30,7 @@ public class EventController {
 
     @PutMapping("/{user}")
     @ResponseStatus(HttpStatus.OK)
-    public void update(@RequestBody Event event,@PathVariable("user") long userId )
+    public void update(@RequestBody Event event,@PathVariable("user") int userId )
     {
         service.update(event,userId);
     }
